feat: add ssh fetch

+++ images/base/git.bash
@@ -69,8 +69,11 @@ gitSshAuth() {
     echo "TARGET_SSH_KEY_PATH ${TARGET_SSH_KEY_PATH} exists, return"
     return
   fi
-  cat /.ssh/id_rsa/ssh-privatekey > "${TARGET_SSH_KEY_PATH}"
-
+  if [ -e /.ssh/id_rsa/ssh-privatekey ]; then
+    cat /.ssh/id_rsa/ssh-privatekey > "${TARGET_SSH_KEY_PATH}"
+  elif [ -e /clusterscanner/git/git_private_key ]; then # same path is used for github private key
+    cat /clusterscanner/git/git_private_key > "${TARGET_SSH_KEY_PATH}"
+  fi
   if [ ! -e ${TARGET_SSH_KEY_PATH} ]; then
     echo "ERROR: Var TARGET_SSH_KEY_PATH is not set, exit"
     exit 1
